stream.py: optionally log keepalive notices
Twitter periodically sends "keepalive" notices to clients connected to
the streaming API.

Before this commit, these notices were silently ignored.

With this commit, every time a keepalive notice is received a
string ("keepalive") is written to sys.stdout if the
TWITTER_LOG_KEEPALIVE environment variable has been set.

This functionality is useful to help monitor the status of streaming
connections.

For example, if you see a log file hasn't received any new tweets in a
few hours but received a keepalive a minute ago, you know the
connection is still active.